Galaxy now uses tool_conf.xml.sample directly
According to the comment in the sample file, if tool_config_files is not
set and tool_conf.xml doesn't exist, Galaxy defaults to using
tool-conf.xml.sample directly. The tool-conf.xml file is not
automatically copied from tool_conf.xml anymore.

Consequently, default tool_config_files to look in tool-conf.xml.sample
instead of tool-conf.xml. This still doesn't reproduce the claimed
default behavior, in that if a tool-conf.xml is created, it will not be
used. Nevertheless, the expectation is that only puppet will be used to
manage the instance, so this is not an issue.

This fixes Galaxy failing to start with:
ConfigurationError: Tool config file not found: ./tool_conf.xml